analysis/GreenSKU-Framework/requirements.txt
ruamel.yaml
numpy
pandas
matplotlib
pathlib
scipy
ipykernel
notebook
tabulate
seaborn
jupyterlab
chardet